+++ backend/modules/location/actions/add.php
@@ -33,7 +33,6 @@ class BackendLocationAdd extends BackendBaseActionAdd
 	{
 		$this->frm = new BackendForm('add');
 		$this->frm->addText('title', null, null, 'inputText title', 'inputTextError title');
-		$this->frm->addEditor('text');
 		$this->frm->addText('street');
 		$this->frm->addText('number');
 		$this->frm->addText('zip');
@@ -62,7 +61,6 @@ class BackendLocationAdd extends BackendBaseActionAdd
 				// build item
 				$item['language'] = BL::getWorkingLanguage();
 				$item['title'] = $this->frm->getField('title')->getValue();
-				$item['text'] = $this->frm->getField('text')->getValue();
 				$item['street'] = $this->frm->getField('street')->getValue();
 				$item['number'] = $this->frm->getField('number')->getValue();
 				$item['zip'] = $this->frm->getField('zip')->getValue();
@@ -76,24 +74,17 @@ class BackendLocationAdd extends BackendBaseActionAdd
 				$item['lng'] = isset($geocode->results[0]->geometry->location->lng) ? $geocode->results[0]->geometry->location->lng : null;
 
 				// insert the item
-				$id = BackendLocationModel::insert($item);
-
-				// add search index
-				// @todo why is this commented out
-				// BackendSearchModel::saveIndex($this->getModule(), (int) $id, array('title' => $item['title'], 'text' => $item['text']));
+				$item['id'] = BackendLocationModel::insert($item);
 
 				// everything is saved, so redirect to the overview
 				if($item['lat'] && $item['lng'])
 				{
 					// trigger event
 					BackendModel::triggerEvent($this->getModule(), 'after_add', array('item' => $item));
-
-					// redirect
-					$this->redirect(BackendModel::createURLForAction('index') . '&report=added&var=' . urlencode($item['title']) . '&highlight=row-' . $id);
 				}
 
-				// could not geocode, redirect to edit
-				else $this->redirect(BackendModel::createURLForAction('edit') . '&id=' . $id);
+				// redirect
+				$this->redirect(BackendModel::createURLForAction('edit') . '&id=' . $item['id'] . '&report=added&var=' . urlencode($item['title']));
 			}
 		}
 	}